Implement groupify using tenant API + added switch for the KeepOldHomePage option
